# services/gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.DeleteUserRequest:
    properties:
      password:
        example: Password123!
        type: string
    required:
    - password
    type: object
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        example: athlete@sportsentral.id
        type: string
    required:
    - email
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        example: athlete@sportsentral.id
        type: string
      password:
        example: Password123!
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponseData:
    properties:
      active_role:
        $ref: '#/definitions/dto.RoleSimpleResponse'
      status:
        $ref: '#/definitions/dto.StatusSimpleResponse'
      tokens:
        $ref: '#/definitions/dto.TokenPairResponse'
      user:
        $ref: '#/definitions/dto.UserSimpleResponse'
    type: object
  dto.LogoutRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
...
  dto.PermissionSimpleResponse:
    properties:
      action:
        type: string
      id:
        type: string
      resource:
        type: string
      slug:
        type: string
    type: object
...
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: athlete@sportsentral.id
        type: string
      full_name:
        example: John Doe
        type: string
      password:
        example: Password123!
        minLength: 8
        type: string
      role_id:
        example: 01950000-0000-7000-8000-000000000001
        type: string
      username:
        example: john_doe
        type: string
    required:
    - email
    - full_name
    - password
    - role_id
    - username
    type: object
...
  dto.ResetPasswordRequest:
    properties:
      password:
        example: NewPassword123!
        minLength: 8
        type: string
      token:
        example: reset-token-uuid
        type: string
    required:
    - password
    - token
    type: object
...
  dto.RoleSimpleResponse:
    properties:
      id:
        type: string
      name:
        type: string
      permission_ids:
        items:
          type: string
        type: array
      permissions:
        items:
          $ref: '#/definitions/dto.PermissionSimpleResponse'
        type: array
      slug:
        type: string
    type: object
...
  dto.SendVerifyEmailRequest:
    properties:
      email:
        example: athlete@sportsentral.id
        type: string
    required:
    - email
    type: object
...
  dto.StatusSimpleResponse:
    properties:
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      type:
        type: string
    type: object
...
  dto.TokenPairResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  dto.UpdateUserRequest:
    properties:
      full_name:
        example: John Doe Updated
        type: string
      username:
        example: john_updated
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      roles:
        items:
          $ref: '#/definitions/dto.UserRoleSimpleResponse'
        type: array
      status:
        $ref: '#/definitions/dto.StatusSimpleResponse'
      updated_at:
        type: string
      username:
        type: string
      verified_at:
        type: string
    type: object
  dto.UserRoleSimpleResponse:
    properties:
      expired_at:
        type: string
      role:
        $ref: '#/definitions/dto.RoleSimpleResponse'
      status:
        $ref: '#/definitions/dto.StatusSimpleResponse'
    type: object
  dto.UserSimpleResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
...
  response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      errors:
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: dev@sportsentral.id
    name: Sport-Sentral Development Team
  description: Centralized API Documentation for Sport-Sentral Microservices Gateway
    (Auth, RBAC, Academy, Venue, Sport, Competition, Scout, Attachment, Log).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sport-Sentral API Gateway
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset instructions & token to registered email.
      parameters:
      - description: Account Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Forgot password request
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login using email & password, returning JWT access & refresh tokens
        and active role.
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponseData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: User authentication & login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke and blacklist the user's refresh token.
      parameters:
      - description: Refresh token to revoke
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Logout from active session
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Obtain a new access token using a valid refresh token.
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokenPairResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new account with an initial role (athlete, scout, court_owner,
        academy_admin).
      parameters:
      - description: User Registration Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register a new user
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using a valid password reset token.
      parameters:
      - description: Password Reset Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Reset password
      tags:
      - Auth
  /auth/verify:
    get:
      description: Verify email and activate account using verification token.
      parameters:
      - description: Account Verification Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Verify user account
      tags:
      - Auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Send account verification link/token to user's email.
      parameters:
      - description: Target Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SendVerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Resend verification email
      tags:
      - Auth
...
  /forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Forgot password (Root alias)
      tags:
      - Users
...
  /me:
    delete:
      consumes:
      - application/json
      description: Soft-delete current user account with password confirmation.
      parameters:
      - description: Password Confirmation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete current user account
      tags:
      - Users
    get:
      description: Retrieve profile data, status, and active role of the currently
        authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get current logged-in user profile
      tags:
      - Auth
    put:
      consumes:
      - application/json
      description: Update username or full name of the currently logged-in account.
      parameters:
      - description: Profile Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - Users
...
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user through the root registration endpoint.
      parameters:
      - description: New User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create new user (Direct register)
      tags:
      - Users
...
  /reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Password Reset Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Reset password (Root alias)
      tags:
      - Users
...
  /verify-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: Target Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SendVerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Send verification email (Root alias)
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token in the format: `Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"